feat: dynamo select mode
